requirements.txt
llama-index==0.11.6
PyMuPDF==1.24.10
python-dotenv
python-box
